Move polybft-secrets command to command package
